feat: complete database support and switch to using i64

+++ src/dto/animal_response_dto.rs
@@ -2,10 +2,10 @@ use serde::{Serialize};
 
 #[derive(Serialize)]
 pub struct AnimalResponseDto {
-    pub id: i32,
+    pub id: i64,
     pub species: String,
     pub common_name: String,
     pub habitat: String,
-    pub lifespan: i32,
+    pub lifespan: i64,
     pub is_endangered: bool
 }
